gitea: include frontend build stage for webpack artifacts
Upstream build strategy changed, meaning the js/css are built and no
longer packaged in the release tarball.

Versions are pinned for go builds and should be for the node
dependency, too.  Using the published node image in the build stage
allows the node and alpine versions to be pinned.

Set up using a build pipeline design, copying the workdir between the
first two stages.

patch included to fix Makefile to allow it to run without the git repo
or drone context

Ref #4